paper-portfolio/pyproject.toml
[project]
name = "paper-portfolio"
version = "0.1.0"
description = "Tools for portfolio construction experiments using datasets and models."
readme = "README.md"
authors = [
    { name = "Lorenzo Varese", email = "55581163+lorenzovarese@users.noreply.github.com" },
]
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.0",
    "polars>=1.30.0",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
]

[project.scripts]
paper-portfolio = "paper_portfolio.run_pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_portfolio"]

[tool.hatch.build.targets.sdist]
include = ["src/paper_portfolio/**", "README.md", "LICENSE"]
